worker-service/requirements.txt
redis==5.0.0
pymongo==4.7.0
python-dotenv==1.0.0
prometheus-client==0.20.0